Added pipeable
Necessary to teach an object how to behave line a pipeline of
transactions so any object that wants to have transaction-like
behavior can do so via the `#pipe` function.

Reek was updated to ignore the nested iterator and too many statement
violations since -- when metaprogramming like this -- this is hard to
avoid without making the code more complex by splitting things up
further. At least all logic is self-contained as best as possible.