Fixed SimpleCov Guard interaction
Necessary to ensure SimpleCov isn't activated when Guard is running.
This prevents SimpleCov reports from being generated when using
focused specs or testing a single implementation because the
corresponding report is only partially built. For a full report, you
must run the entire test suite which is perfect for Continous
Integration use as well.