Added benchmarks
Necessary to inspect the performance of this gem compared against the
Dry Transaction gem since both gems are similar in nature. For the
moment, there is a distinctive performance hit when using the *full
features* of Transactable which needs to be addressed. When only using
composition (without the abstract class and instrumentation) speeds
significantly increase. There is work to do before getting to Version
1.0.0.

I've ignored the Reek violations -- which are legit and should
normally not be used -- because I needed to benchmark this gem without
relying on native instrumentation support.