Updated to Infusible 3.1.0
Necessary to leverage the new _infused keys_ feature which allows us to access the keys of all infused dependencies which removes the need for the constant. We still have to define the dependencies a second time when passing to `super` but then have direct access to `infused_keys` after which is nice. link:https://alchemists.io/projects/infusible/versions[Details].

Milestone: minor